Move TreeNode to model (from model.group)
