+++ src/test/java/net/sf/jabref/model/TreeNodeTest.java
@@ -1,4 +1,4 @@
-package net.sf.jabref.model.groups;
+package net.sf.jabref.model;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Consumer;
 
+import net.sf.jabref.model.TreeNode;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
